+++ docs/change-log.md
@@ -13,6 +13,7 @@ outline: deep
 * Implemented: figure out a way to share file extensions for s3files from schema #104
 * NOTE: Because of the above you may want to check out the new acceptedFileExtensions attribute in ving schemas and migrate your S3File integrations to use it.
 * Implemented: add a display of an s3file thumbnail to the page generator #105
+* Fixed: pulumi doesn't create the nodmods.zip file as it should #110
 
 ## 2024-04-23
 * Created SelectInput component to replace FormSelect. However, you should use FormInput with type select instead of using this directly in most cases.